backend/requirements.txt
django=5.1.3
django-cors-headers=4.6.0
django-dotenv=1.4.2
djangorestframework=3.15.2
djangorestframework-simplejwt=5.3.1
